psq-server: don't apply default on read error
Especially when dialing up log levels, it can be easily missed if the
explicitly specified config file does not exist.

Instead of having a String with a default path, make it an Option<_>, so
we know later if the user explicitly specified a config or not.

In case they didn't, we can fallback to Config::create_default().

Also, make the types a bit more idiomatic - use a PathBuf, and make
Config::read_from_file take any AsRef<Path>, which includes &str, &Path,
PathBuf and String.